requirements.txt
gql
aiohttp
slack-sdk
python-dotenv
influxdb3-python
